# packages/api-contracts/openapi/trademesh-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeMesh API",
    "description": "Reviewed contract for the TradeMesh web client and partner integrations.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "Current TradeMesh API host" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/tracking/{shipmentId}/position": {
      "post": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryPosition",
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-Telemetry-Credential",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PositionRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PositionIngestionResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        },
        "security": []
      }
    },
    "/api/telemetry/readings": {
      "post": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryIngest",
        "parameters": [
          {
            "name": "X-Telemetry-Credential",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IngestReadingsRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/IngestionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        },
        "security": []
      }
    },
...
    "/api/businesses/{businessId}/shipments/{shipmentId}/telemetry-devices": {
      "post": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryProvision",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedDeviceResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
    "/api/tracking/{shipmentId}/events": {
      "get": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryPositionEvents",
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "businessId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": { "$ref": "#/components/schemas/SseEmitter" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/tracking/{shipmentId}/backhaul-matches": {
      "get": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryBackhaulMatches",
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "businessId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BackhaulMatchesResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
    "/api/businesses/{businessId}/shipments/{shipmentId}/telemetry": {
      "get": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryHistory",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingHistoryResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
    "/api/businesses/{businessId}/telemetry-devices/{deviceId}": {
      "delete": {
        "tags": ["telemetry-controller"],
        "operationId": "telemetryRevoke",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PositionRequest": {
        "type": "object",
        "properties": {
          "clientEventId": { "type": "string", "format": "uuid" },
          "recordedAt": { "type": "string", "format": "date-time" },
          "lat": {
            "type": "number",
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0
          },
          "lng": {
            "type": "number",
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0
          },
          "speedKilometresPerHour": { "type": "number" },
          "batteryPercent": { "type": "number" },
          "networkStatus": {
            "type": "string",
            "enum": ["CONNECTED", "LIMITED", "OFFLINE", "UNKNOWN"]
          },
          "networkSignalDbm": { "type": "integer", "format": "int32" }
        },
        "required": ["lat", "lng"]
      },
      "PositionIngestionResponse": {
        "type": "object",
        "properties": {
          "shipmentId": { "type": "string", "format": "uuid" },
          "clientEventId": { "type": "string", "format": "uuid" },
          "readingId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["ACCEPTED", "DUPLICATE"] },
          "recordedAt": { "type": "string", "format": "date-time" },
          "latitude": { "type": "number", "format": "double" },
          "longitude": { "type": "number", "format": "double" },
          "speedKilometresPerHour": { "type": "number" }
        }
      },
      "IngestReadingsRequest": {
        "type": "object",
        "properties": {
          "readings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReadingRequest" },
            "minItems": 1
          }
        },
        "required": ["readings"]
      },
      "ReadingRequest": {
        "type": "object",
        "properties": {
          "clientEventId": { "type": "string", "format": "uuid" },
          "recordedAt": { "type": "string", "format": "date-time" },
          "latitude": { "type": "number", "format": "double" },
          "longitude": { "type": "number", "format": "double" },
          "speedKilometresPerHour": { "type": "number" },
          "fuelLitres": { "type": "number" },
          "temperatureCelsius": { "type": "number" },
          "sealOpen": { "type": "boolean" },
          "batteryPercent": { "type": "number" },
          "networkStatus": {
            "type": "string",
            "enum": ["CONNECTED", "LIMITED", "OFFLINE", "UNKNOWN"]
          },
          "networkSignalDbm": { "type": "integer", "format": "int32" }
        },
        "required": ["clientEventId", "recordedAt"]
      },
      "IngestionResponse": {
        "type": "object",
        "properties": {
          "readings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReadingReceiptResponse" }
          },
          "acceptedCount": { "type": "integer", "format": "int32" },
          "duplicateCount": { "type": "integer", "format": "int32" }
        }
      },
      "ReadingReceiptResponse": {
        "type": "object",
        "properties": {
          "clientEventId": { "type": "string", "format": "uuid" },
          "readingId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["ACCEPTED", "DUPLICATE"] }
        }
      },
...
      "ProvisionDeviceRequest": {
        "type": "object",
        "properties": {
          "displayName": { "type": "string", "maxLength": 120, "minLength": 0 }
        },
        "required": ["displayName"]
      },
      "IssuedDeviceResponse": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "string", "format": "uuid" },
          "shipmentId": { "type": "string", "format": "uuid" },
          "displayName": { "type": "string" },
          "status": { "type": "string", "enum": ["ACTIVE", "REVOKED"] },
          "credential": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "PointResponse": {
        "type": "object",
        "properties": {
          "label": { "type": "string" },
          "latitude": { "type": "number", "format": "double" },
          "longitude": { "type": "number", "format": "double" }
        }
      },
...
      "SseEmitter": {
        "type": "object",
        "properties": { "timeout": { "type": "integer", "format": "int64" } }
      },
      "BackhaulMatchResponse": {
        "type": "object",
        "properties": {
          "shipmentId": { "type": "string", "format": "uuid" },
          "businessId": { "type": "string", "format": "uuid" },
          "pickup": { "$ref": "#/components/schemas/PointResponse" },
          "destination": { "$ref": "#/components/schemas/PointResponse" },
          "windowStart": { "type": "string", "format": "date-time" },
          "windowEnd": { "type": "string", "format": "date-time" },
          "pickupDistanceMetres": { "type": "integer", "format": "int64" },
          "pickupDurationSeconds": { "type": "integer", "format": "int64" },
          "roadDistanceMeasured": { "type": "boolean" },
          "averageRating": { "type": "number" },
          "successfulDeliveryRate": { "type": "number" },
          "trustScore": { "type": "number" },
          "score": { "type": "number" }
        }
      },
      "BackhaulMatchesResponse": {
        "type": "object",
        "properties": {
          "matches": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackhaulMatchResponse" }
          }
        }
      },
...
      "ReadingHistoryResponse": {
        "type": "object",
        "properties": {
          "readings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReadingResponse" }
          },
          "units": { "$ref": "#/components/schemas/UnitsResponse" }
        }
      },
      "ReadingResponse": {
        "type": "object",
        "properties": {
          "readingId": { "type": "string", "format": "uuid" },
          "deviceId": { "type": "string", "format": "uuid" },
          "clientEventId": { "type": "string", "format": "uuid" },
          "recordedAt": { "type": "string", "format": "date-time" },
          "receivedAt": { "type": "string", "format": "date-time" },
          "latitude": { "type": "number", "format": "double" },
          "longitude": { "type": "number", "format": "double" },
          "speedKilometresPerHour": { "type": "number" },
          "fuelLitres": { "type": "number" },
          "temperatureCelsius": { "type": "number" },
          "sealOpen": { "type": "boolean" },
          "batteryPercent": { "type": "number" },
          "networkStatus": {
            "type": "string",
            "enum": ["CONNECTED", "LIMITED", "OFFLINE", "UNKNOWN"]
          },
          "networkSignalDbm": { "type": "integer", "format": "int32" },
          "retentionTier": { "type": "string", "enum": ["RAW", "DOWNSAMPLED"] }
        }
      },
      "UnitsResponse": {
        "type": "object",
        "properties": {
          "coordinates": { "type": "string" },
          "speed": { "type": "string" },
          "fuel": { "type": "string" },
          "temperature": { "type": "string" },
          "battery": { "type": "string" },
          "networkSignal": { "type": "string" }
        }
      },
...
      "ApiProblem": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "example": "about:blank" },
          "title": { "type": "string", "example": "The request is invalid" },
          "status": { "type": "integer", "format": "int32", "example": 400 },
          "detail": {
            "type": "string",
            "example": "One or more fields are invalid."
          },
          "instance": { "type": "string", "example": "/api/auth/login" },
          "code": { "type": "string", "example": "INVALID_REQUEST" },
          "requestId": {
            "type": "string",
            "example": "01JEXAMPLE00000000000000000"
          }
        },
        "required": [
          "code",
          "detail",
          "instance",
          "requestId",
          "status",
          "title",
          "type"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 400,
              "instance": "/api/example",
              "code": "INVALID_REQUEST",
              "detail": "The request is invalid.",
              "title": "The request is invalid",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 401,
              "instance": "/api/example",
              "code": "AUTHENTICATION_REQUIRED",
              "detail": "Authentication is required.",
              "title": "Authentication is required",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The caller is not allowed to perform this action",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 403,
              "instance": "/api/example",
              "code": "ACCESS_DENIED",
              "detail": "The caller is not allowed to perform this action.",
              "title": "The caller is not allowed to perform this action",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 404,
              "instance": "/api/example",
              "code": "NOT_FOUND",
              "detail": "The requested resource was not found.",
              "title": "The requested resource was not found",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with current state",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 409,
              "instance": "/api/example",
              "code": "STATE_CONFLICT",
              "detail": "The request conflicts with current state.",
              "title": "The request conflicts with current state",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The request limit was exceeded",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 429,
              "instance": "/api/example",
              "code": "RATE_LIMIT_EXCEEDED",
              "detail": "The request limit was exceeded.",
              "title": "The request limit was exceeded",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "ServerError": {
        "description": "The server could not complete the request",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 500,
              "instance": "/api/example",
              "code": "INTERNAL_ERROR",
              "detail": "The server could not complete the request.",
              "title": "The server could not complete the request",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "BadGateway": {
        "description": "An external provider rejected the request",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 502,
              "instance": "/api/example",
              "code": "EXTERNAL_PROVIDER_FAILED",
              "detail": "An external provider rejected the request.",
              "title": "An external provider rejected the request",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "An external provider is temporarily unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "about:blank",
              "status": 503,
              "instance": "/api/example",
              "code": "EXTERNAL_PROVIDER_UNAVAILABLE",
              "detail": "An external provider is temporarily unavailable.",
              "title": "An external provider is temporarily unavailable",
              "requestId": "01JEXAMPLE00000000000000000"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}